services/tuner-controller/src/main.py
import time
import logging
import os
import requests
import json
import yaml
import numpy as np
print(f"DEBUG: NumPy Version: {np.__version__}")
# Patch for scikit-optimize compatibility with newer numpy
if not hasattr(np, 'int'):
    print("DEBUG: Patching np.int")
    np.int = int
else:
    print("DEBUG: np.int already exists")

from prometheus_api_client import PrometheusConnect
from skopt import Optimizer
from skopt.space import Integer
from kubernetes import client, config as k8s_config
from scipy import stats

# Configure Logging
logging.basicConfig(level=logging.INFO, format='%(asctime)s %(levelname)s %(message)s')
logger = logging.getLogger(__name__)

GATE_URL = os.getenv("GATE_URL", "http://constraint-gate:8000")
PROMETHEUS_URL = os.getenv("PROMETHEUS_URL", "http://prometheus:9090")
ROLLOUT_MANAGER_URL = os.getenv("ROLLOUT_MANAGER_URL", "http://rollout-manager:8000")
NAMESPACE = os.getenv("NAMESPACE", "default")
SCENARIO_PATH = os.getenv("SCENARIO_PATH", "/app/configs/scenarios/S1_steady.yaml")

class TunerSystem:
    def __init__(self1, slos):
        self1.slos = slos
        # Search space: Parallelism [1, 8], Kafka Batch Size [1k, 5MB]
        self1.space = [
            Integer(1, 8, name='parallelism'),
            Integer(4096, 5242880, name='batch_size')
        ]
        self1.optimizer = Optimizer(self1.space, base_estimator="GP", acq_func="EI")
        self1.default_config = {
            "flink": {
                "parallelism": 2,
                "checkpoint_interval_ms": 60000,
                "checkpoint_timeout_ms": 120000,
                "unaligned_checkpoints": False,
                "max_concurrent_checkpoints": 1,
                "state_ttl_ms": 300000,
                "managed_memory_mb": 256,
                "state_backend": "hashmap"
            },
            "kafka": {
                "batch_size": 16384,
                "partitions": 4,
                "min_insync_replicas": 1,
                "replication_factor": 1,
                "retention_ms": 600000
            }
        }
        self1.current_config = self1.default_config['flink']
        self1.history = []
        
        # K8s Setup
        try:
            try:
                k8s_config.load_incluster_config()
            except:
                k8s_config.load_kube_config()
            self1.apps_v1 = client.AppsV1Api()
            logger.info("✓ Kubernetes API initialized.")
        except Exception as e:
            logger.warning(f"Kubernetes API NOT initialized: {e}. Orchestration will rely on simulated rollout.")
            self1.apps_v1 = None
        self1.refresh_current_config()

    def refresh_current_config(self1):
        """Fetch actual state from K8s to avoid hardcoded assumptions."""
        logger.info("Syncing with live K8s configuration...")
        try:
            deploy = self1.apps_v1.read_namespaced_deployment(name="flink-taskmanager", namespace=NAMESPACE)
            live_parallelism = deploy.spec.replicas
            self1.default_config['flink']['parallelism'] = live_parallelism
            self1.current_config = self1.default_config['flink']
            logger.info(f"✓ Synced: current parallelism is {live_parallelism}")
        except Exception as e:
            logger.warning(f"Failed to sync with K8s: {e}. Falling back to defaults.")
            self1.current_config = self1.default_config['flink']

    def get_metrics(self1, prom):
        try:
            # 1. Throughput (Out)
            throughput_data = prom.custom_query(query="sum(flink_taskmanager_job_task_operator_numRecordsOutPerSecond)")
            throughput = float(throughput_data[0]['value'][1]) if throughput_data else 0.0

            # 2. Consumer Lag
            lag_data = prom.custom_query(query="sum(kafka_consumergroup_lag{consumergroup='flink-consumer'})")
            lag = float(lag_data[0]['value'][1]) if lag_data else 0.0

            # 3. CPU Usage
            cpu_data = prom.custom_query(query="sum(flink_taskmanager_Status_JVM_CPU_Load)")
            cpu = float(cpu_data[0]['value'][1]) if cpu_data else 0.0
            
            # Latency (ms) - Using real heuristic from paper but based on live metrics
            latency = (lag / (throughput + 1)) * 1000 if throughput > 0 else 0
            
            return {"throughput": throughput, "lag": lag, "cpu": cpu, "latency": latency}
        except Exception as e:
            logger.error(f"Error fetching metrics: {e}")
            return None

    def perform_statistical_test(self1, baseline_metrics, canary_metrics):
        """Wilcoxon signed-rank test and 95% Bootstrap CI."""
        # In a real system, we'd collect multiple samples over the canary window
        # For this implementation, we simulate the paired test logic described in the paper
        b_vals = [m['latency'] for m in baseline_metrics]
        c_vals = [m['latency'] for m in canary_metrics]
        
        if len(b_vals) < 5: return True # Not enough data, assume OK for demo
        
        stat, p_val = stats.wilcoxon(b_vals, c_vals)
        return p_val < 0.05

    def apply_config(self1, cfg):
        logger.info(f"Applying config to K8s: {cfg}")
        try:
            # Patch TaskManager replicas for parallelism
            body = {"spec": {"replicas": cfg['parallelism']}}
            self1.apps_v1.patch_namespaced_deployment(
                name="flink-taskmanager",
                namespace=NAMESPACE,
                body=body
            )
            # In a real system we'd also trigger a Flink job update for batch_size
            self1.current_config = cfg
            return True
        except Exception as e:
            logger.error(f"K8s Patch failed: {e}")
            return False

    def evaluate_cost(self1, metrics):
        if not metrics: return 10000
        
        lat = metrics['latency']
        target = self1.slos.get('latency_p95_ms', 3000)
        
        # Penalize SLO breaches heavily
        if lat > target:
            return 1000 + (lat - target)
        
        # Optimize for resource efficiency (fewer cores) while staying under target
        return metrics['cpu'] * 10 

    def step(self1, prom):
        # 1. Observe
        metrics = self1.get_metrics(prom)
        if metrics:
            cost = self1.evaluate_cost(metrics)
            # 2. Tell optimizer about last result (if we just applied a new one)
            # For simplicity, we assume steady state reached
            logger.info(f"Iteration metrics: {metrics}, Cost: {cost:.2f}")
            
        # 3. Ask for next candidate
        while True:
            candidate_vec = self1.optimizer.ask()
            
            # Build full candidate by deep merging or simply copying default
            import copy
            candidate = copy.deepcopy(self1.default_config)
            candidate['flink']['parallelism'] = int(candidate_vec[0])
            candidate['kafka']['batch_size'] = int(candidate_vec[1])
            
            # 4. Initiate Staged Rollout via Rollout Manager
            logger.info(f"Initiating Rollout through {ROLLOUT_MANAGER_URL}")
            try:
                payload = {
                    "candidate": candidate,
                    "namespace": NAMESPACE
                }
                resp = requests.post(f"{ROLLOUT_MANAGER_URL}/rollout", json=payload, timeout=600)
                result = resp.json()
                
                if resp.status_code == 200 and result.get('status') == 'success':
                    logger.info(f"✓ Rollout SUCCESSFUL: {result.get('message')}")
                    # Update local state
                    self1.current_config = candidate['flink']
                    
                    # Update optimizer
                    last_vec = [candidate['flink']['parallelism'], candidate['kafka']['batch_size']]
                    self1.optimizer.tell(last_vec, 100) 
                    
                    record = {
                        "timestamp": time.time(),
                        "candidate": candidate,
                        "action": "rolled_out",
                        "p_value": result.get('p_value')
                    }
                    self1.write_record(record)
                    break # Success, move to next iteration
                else:
                    logger.warning(f"✗ Rollout REJECTED/FAILED in phase {result.get('phase')}: {result.get('message')}")
                    # Penalize in optimizer
                    self1.optimizer.tell(candidate_vec, 5000)
                    record = {
                        "timestamp": time.time(),
                        "candidate": candidate,
                        "action": f"rejected_{result.get('phase')}",
                        "reason": result.get('message')
                    }
                    self1.write_record(record)
                    # Loop will ask for next candidate
            except Exception as e:
                logger.error(f"Rollout Manager error: {e}")
                time.sleep(5)

    def write_record(self1, record):
        try:
             with open("/app/artifacts/decisions.jsonl", "a") as f:
                f.write(json.dumps(record) + "\n")
        except: pass

def main():
    logger.info("Starting REAL Tuner Controller (Bayesian Optimization)...")
    prom = PrometheusConnect(url=PROMETHEUS_URL, disable_ssl=True)
    
    # Wait for Prometheus
    for _ in range(10):
        if prom.check_prometheus_connection():
            logger.info("Connected to Prometheus.")
            break
        logger.info("Waiting for Prometheus...")
        time.sleep(5)

    with open(SCENARIO_PATH, 'r') as f:
        scenario = yaml.safe_load(f)
    
    slos = scenario.get('slos', {})
    tuner = TunerSystem(slos)

    while True:
        try:
            tuner.step(prom)
        except Exception as e:
            logger.error(f"Error in control loop: {e}")
        time.sleep(60) # Measurement window

if __name__ == "__main__":
    main()
