services/tuner-controller/requirements.txt
requests==2.31.0
prometheus-api-client==0.5.3
kubernetes==28.1.0
pyyaml==6.0.1
scikit-optimize==0.9.0
numpy==1.23.5
scipy==1.11.2
